backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
openai==2.29.0
python-multipart==0.0.22
pillow==12.1.1
torch==2.11.0
torchvision==0.26.0
transformers==5.3.0
faiss-cpu==1.13.2
pydantic==2.12.5
python-dotenv==1.2.2
aiofiles==25.1.0
mem0ai==1.0.7
